EDIT: api 종류에 따른 patient.js / doctor.js 분리
